Move explode and list logic to lib directory
